# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flock Energy - Urja Meter Ops API",
    "description": "\nA clean, documented REST API over the legacy **Urja Meter Ops** portal.\n\nThe portal has no API. This service logs in as a normal user, manages the session,\nreverse-engineers the portal's internal endpoints, normalises its inconsistent payloads,\nand exposes the result as typed JSON that another program can consume without ever\ntouching the portal.\n\n**What it adds over the portal**\n\n* Filtering, sorting and pagination the portal cannot do (make, status, phase, build,\n  any hierarchy code).\n* The network hierarchy, reconstructed from per-meter paths into a queryable tree.\n* Real consumption, derived by differencing the portal's cumulative registers.\n* Proximity search over meter coordinates.\n* An explicit data-quality report instead of silently tidied-away inconsistencies.\n\nSee `PROTOCOL.md` for how the portal actually works, and `README.md` for design decisions\nand trade-offs.\n",
    "contact": {
      "name": "Flock Energy take-home submission"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/meters": {
      "get": {
        "tags": [
          "Meters"
        ],
        "summary": "List meters",
        "description": "Filter, sort and page the full meter estate.\n\nServed from the local snapshot, so this supports filters the portal itself cannot: by make, status, phase, install type, build and any hierarchy code. All filters are ANDed. `search` matches meter id or serial as a literal substring - unlike the portal, `%` and `_` are not wildcards here.",
        "operationId": "list_meters_api_v1_meters_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Literal substring of meter id or serial.",
              "title": "Search"
            },
            "description": "Literal substring of meter id or serial."
          },
          {
            "name": "make",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact manufacturer, e.g. 'HPL'.",
              "title": "Make"
            },
            "description": "Exact manufacturer, e.g. 'HPL'."
          },
          {
            "name": "install_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "installed | faulty | decommissioned",
              "title": "Install Status"
            },
            "description": "installed | faulty | decommissioned"
          },
          {
            "name": "phase_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "single | three",
              "title": "Phase Type"
            },
            "description": "single | three"
          },
          {
            "name": "install_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "whole_current | ct_operated",
              "title": "Install Type"
            },
            "description": "whole_current | ct_operated"
          },
          {
            "name": "build",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "legacy | v2",
              "title": "Build"
            },
            "description": "legacy | v2"
          },
          {
            "name": "dt_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Distribution transformer code, e.g. 'DT-001'.",
              "title": "Dt Code"
            },
            "description": "Distribution transformer code, e.g. 'DT-001'."
          },
          {
            "name": "hierarchy_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Any hierarchy code at any level, e.g. 'F-002' or 'Z-01'.",
              "title": "Hierarchy Code"
            },
            "description": "Any hierarchy code at any level, e.g. 'F-002' or 'Z-01'."
          },
          {
            "name": "has_location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only meters with/without coordinates.",
              "title": "Has Location"
            },
            "description": "Only meters with/without coordinates."
          },
          {
            "name": "with_issues",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only meters with/without data-quality flags.",
              "title": "With Issues"
            },
            "description": "Only meters with/without data-quality flags."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "One of: meter_id, serial_no, make, install_status, dt_code",
              "default": "meter_id",
              "title": "Sort"
            },
            "description": "One of: meter_id, serial_no, make, install_status, dt_code"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "asc",
              "title": "Order"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-based page number.",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page number."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Items per page (max 200).",
              "default": 25,
              "title": "Page Size"
            },
            "description": "Items per page (max 200)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MeterSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/near": {
      "get": {
        "tags": [
          "Meters"
        ],
        "summary": "Find meters near a location",
        "description": "Great-circle proximity search over the snapshot - a question the portal cannot answer at all, since it never exposes coordinates in its listing.\n\nDeclared before `/{meter_id}` so the literal path wins over the parameter.",
        "operationId": "meters_near_point_api_v1_meters_near_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "description": "Latitude in decimal degrees.",
              "title": "Lat"
            },
            "description": "Latitude in decimal degrees."
          },
          {
            "name": "lng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "description": "Longitude in decimal degrees.",
              "title": "Lng"
            },
            "description": "Longitude in decimal degrees."
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 500,
              "exclusiveMinimum": 0,
              "description": "Search radius in kilometres.",
              "default": 1.0,
              "title": "Radius Km"
            },
            "description": "Search radius in kilometres."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MeterWithDistance"
                  },
                  "title": "Response Meters Near Point Api V1 Meters Near Get"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}": {
      "get": {
        "tags": [
          "Meters"
        ],
        "summary": "Get one meter",
        "description": "Full record including the normalised root-to-leaf hierarchy path and coordinates.\n\nUpstream this data has two incompatible encodings depending on the meter's build; both are normalised to this single shape.",
        "operationId": "get_meter_api_v1_meters__meter_id__get",
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Case-sensitive portal id, e.g. 'J100001'.",
              "title": "Meter Id"
            },
            "description": "Case-sensitive portal id, e.g. 'J100001'."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meter"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such meter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}/consumption": {
      "get": {
        "tags": [
          "Meters"
        ],
        "summary": "Get consumption history",
        "description": "Energy **consumed**, derived by differencing the portal's cumulative registers.\n\nThe portal returns raw register totals; summing that column would overstate usage by orders of magnitude. Each interval here is the difference between consecutive readings. Sampling cadence is inferred from the data (30-minute for most meters, daily for some), never assumed.\n\nThe upstream series is a fixed ~7-day window and ignores date parameters, so `from`/`to` are applied locally.",
        "operationId": "get_consumption_api_v1_meters__meter_id__consumption_get",
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Case-sensitive portal id.",
              "title": "Meter Id"
            },
            "description": "Case-sensitive portal id."
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive lower bound (ISO-8601, naive local time).",
              "title": "From"
            },
            "description": "Inclusive lower bound (ISO-8601, naive local time)."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive upper bound (ISO-8601, naive local time).",
              "title": "To"
            },
            "description": "Inclusive upper bound (ISO-8601, naive local time)."
          },
          {
            "name": "granularity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "raw | hourly | daily",
              "default": "raw",
              "title": "Granularity"
            },
            "description": "raw | hourly | daily"
          },
          {
            "name": "include_readings",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Also return the raw cumulative register readings.",
              "default": false,
              "title": "Include Readings"
            },
            "description": "Also return the raw cumulative register readings."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such meter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/hierarchy/nodes/{node_id}/meters": {
      "get": {
        "tags": [
          "Hierarchy"
        ],
        "summary": "List meters under a subtree",
        "description": "Every meter whose path passes through the given node, paginated.",
        "operationId": "get_node_meters_api_v1_hierarchy_nodes__node_id__meters_get",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Path id, e.g. 'Z-01/C-01'.",
              "title": "Node Id"
            },
            "description": "Path id, e.g. 'Z-01/C-01'."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-based page number.",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page number."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Items per page (max 200).",
              "default": 25,
              "title": "Page Size"
            },
            "description": "Items per page (max 200)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MeterSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No node with that path id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Envelope returned for **every** non-2xx response.\n\nDeclared so the OpenAPI document describes the real contract. Without it FastAPI\nadvertises its own ``HTTPValidationError`` for 422 and nothing at all for 404/5xx,\nwhich makes generated clients fail to deserialise every error they receive."
      },
...
      "Page_MeterSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MeterSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "Page[MeterSummary]"
      },
...
    }
  },
...
}